fix(diagnostics): correct path matching behavior and add auto-truncation
- Fix path matcher to return false for empty pattern instead of true
- Introduce auto-truncation logic for low cardinality metrics to replace variant parameters (e.g., actor IDs) with placeholders
- Refactor HTTP monitoring to separate path matching and auto-truncation concerns
- Move path matching tests to dedicated file and add comprehensive test coverage
